storage: move requestLeaderLease to its own file
To be refactored next.
